Do not add retention lease inside ShardReplicationTask
