+++ src/main/kotlin/org/opensearch/replication/task/index/IndexReplicationTask.kt
@@ -123,7 +123,6 @@ open class IndexReplicationTask(id: Long, type: String, action: String, descript
 
     override val log = Loggers.getLogger(javaClass, Index(params.followerIndexName, ClusterState.UNKNOWN_UUID))
     private val retentionLeaseHelper = RemoteClusterRetentionLeaseHelper(clusterService.clusterName.value(), remoteClient)
-
     private var shouldCallEvalMonitoring = true
     private var updateSettingsContinuousFailCount = 0
     private var updateAliasContinousFailCount = 0
